Pass the current version to the search
From a UX POV, when I'm searching docs I'm likely to be working on a site of a specific version - results of docs from older or newer versions are likely to be less relevant to me. If I want to search wider after my initial search I can use the check boxes on the advanced search.