FIX horizontal alignment shifts when vertical scroll bar appears
Common problem. 

Quite visible if (i) a page's content fits within the browser window's height when the table of contents dropdown is closed but (ii) the content no longer fits vertically when the table of contents dropdown is opened. When this happens, a vertical scroll bar appears in the browser and the page's horizontal alignment shifts by the scroll bar width.  

@frankmullenger I also noticed this in swipestripe docs :)